Update documentation for ENA DPDK driver
The README file was changed to README.md due to markdown formatting
being used there.

README.md was updates with instructions on how to use vfio-pci module
with the ENA on AWS.

Release noted were added, so all changes for the ENA PMD can be tracked
together with the DPDK relases (including LTS releases).