linux/efa: Add spec file for EFA kernel driver
Add a spec file to package the EFA kernel driver source and install it
using dkms. This is so we can install the driver in an environment where
we do not control the kernel installation.  We should add another spec
file to build the kernel driver and package it when we package the
driver with the kernel.

Signed-off-by: Robert Wespetal <wesper@amazon.com>
Signed-off-by: Gal Pressman <galpress@amazon.com>