linux/efa: Remove rdma_user_mmap_page usage
The rdma_user_mmap_page function has been removed from the kernel and
vm_insert_page should be used instead.
The VMA holds a reference on the page and it will be automatically freed
when the last reference drops. No need for disassociate to sequence the
cleanup.

The vm_flags checks that were previously done by the function are now
done explicitly by the driver:
- Make sure the pages are shared
- Make sure the pages are not executable
- Make sure mprotect doesn't make the pages executable

Reviewed-by: Firas JahJah <firasj@amazon.com>
Reviewed-by: Yossi Leybovich <sleybo@amazon.com>
Signed-off-by: Gal Pressman <galpress@amazon.com>