dpdk/ena: update release notes up to v19.08
As the v19.08 is out for some time, the release notes of the PMD
were updated.

Also, the information about stable releases was added.

In the v16.04 there was misleading information that ENA there is r1.0.0,
while the version r1.0.0 was added in the DPDK v16.07 release.
To avoid misunderstandings, the ENA in v16.04 version is now being
called "unversioned", while the one in the v16.07 is now r1.0.0.

Signed-off-by: Michal Krawczyk <mk@semihalf.com>