linux/efa: Refactor GDR locking scheme
Current GDR locking scheme depends on acquiring the EFA nvmem lock for
the entire flow of reg/dereg_mr and free callback, thus guaranteeing
"atomicity" of nvmem and MRs, where either both exist or none.

This scheme could lead to potential deadlock due to circular locking in
the form of:

Process A                                    Process B
---------                                    ---------
register MR
acquire nvmem lock                           acquire nvidia internal lock
  nvmem_get                                    free callback
    nvidia_p2p_get_pages                         acquire nvmem lock
      acquire nvidia internal lock

This issue has been observed on newer version of NCCL, where the library
validates GPUDirect RDMA support by issuing a reg MR on CUDA buffers,
tests whether it succeeded or not and deregisters it immediately, which
might result in the free callback triggering while other processes are
still in their registration (i.e nvidia_p2p_get_pages) part. This is as
opposed to older version of NCCL where the pattern was usually: memory
registrations during initialization -> run some traffic ->
deregisterations during cleanup.

In order to solve that we make a few changes to our locking scheme.
Mainly, we no longer hold the nvmem mutex while calling nvidia's get
APIs and the nvmem lock is only acquired for a small set of operations.
This change assumes that the user is sane, and does not fork his
process/uses threads to free the buffers before the MR is registered,
and does not try to free the buffers in parallel to a dereg operation.
Doing so will most likely result in a backtrace which is used to
identify incorrect behavior.

Reviewed-by: Rashika Kheria <rashika@amazon.com>
Reviewed-by: Yossi Leybovich <sleybo@amazon.com>
Signed-off-by: Gal Pressman <galpress@amazon.com>