dpdk: update script for adding WC to the vfio
Previous patch was hardcoded and had some issues with compatibility with
few kernel versions.

To avoid that, the patch is now kept in separate file and can be choosen
dynamically using the kernel version.

As for now, Linux v4.10 and all later are supported.

Besides that, the script now prints out much more useful messages, have
improved error reporting and error handling and bash style has been
unified.

The REAME for the script was updated as well - including list of AMIs
supported by the script, few MD style fixes and other minor changes.

Signed-off-by: Michal Krawczyk <mk@semihalf.com>