linux/ena: Propagate upstream support for AF XDP busypoll
AF XDP RX/TX code path is different from the code path for non-xdp
sockets. As a result, the napi id, which is passed inside sk_buff struct
for non-xdp flows, is not available for AF XDP sockets.

The napi id is an identifier of the napi struct used by the driver in
its polling function ena_io_poll() and is needed by busy poll routine which
calls this function repeatedly.

To accommodate for the missing napi id patch [1] passes it as part of
the xdp_rxq_info struct when loading an eBPF program to an interface.
Passing 0 as the napi id indicates that busy poll for AF XDP is not
supported for the driver (which is indeed the case for the ENA driver).
The full patchset which adds the busy poll support can be found in [2].

This patch propagates the upstream change to the driver.

[1]: https://www.spinics.net/lists/netdev/msg704553.html
[2]: https://www.spinics.net/lists/netdev/msg704547.html

Signed-off-by: Shay Agroskin <shayagr@amazon.com>
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>