linux/ena: Change ena_increase_stat_atomic() function name
The function was introduced in commit [1] in order to aggregate the
ubiquitous three lines
    u64_stats_update_begin(syncp);
    *pointer_to_stat += value;
    u64_stats_update_end(syncp);

The function was called ena_increase_stat_atomic().

During upstream review [2] we got a comment that the name might be
misleading since on 64bit architectures the lines above are translated
into:
    *pointer_to_stat += value;

This is because 64bit STORE ops on these architectures are done
atomically.

[1]: See 'Fixes' tag
[2]: https://lore.kernel.org/netdev/4bfce37b-674e-b738-46a6-d4d8af6ac8cb@gmail.com/T/#u

Fixes: d6abd61df5ee ("linux/ena update ENA linux driver to version 2.3.0")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>