dpdk/ena: backport ENA v2.4.0 for the LTS releases
Instead of supplying users with only the critical fixes, the whole ENA
driver was backported for the supported LTS branches.

ENA features which are depending on the DPDK API missing in the targeted
release were not backported, and in order to use them, upgrade to the
newer DPDK version is needed.

The backports were created for given LTS releases:

* v16.11.11
* v17.11.10
* v18.11.11
* v19.11.10
* v20.11.3

Folders containing critical ENA fixes for the main DPDK releases were
removed, and instead the ENA backports based on the LTS branches can be
used. The patches shouldn't be applied to the DPDK version other than
the one on which the ENA is based. Otherwise, there can be multiple
merge conflicts or even ENA may end up with some important patches being
missing if they were already a part of the LTS release.

For convenience, the `userspace/dpdk/backports/apply-patches.sh` script
can be used. Alternatively, the patches can be applied manually.

Signed-off-by: Michal Krawczyk <mk@semihalf.com>