fbsd/ena: convert README format to reStructuredText
Instead of using raw txt, the reStructuredText format was added, to make
the README file easier to read. The similar change was made to the Linux
kernel driver [1].

This change fixes some typos, changes the basic structure of the file
and adds minor missing information (namely about FreeBSD 14 support)

Reference to '2.0.0' driver was changed to '2.3.1'.

[1] https://lkml.org/lkml/2020/5/1/848

Signed-off-by: Michal Krawczyk <mk@semihalf.com>