fbsd/ena: update README steps for getting kernel sources
As FreeBSD project moved to Git and dropped the subversion support, the
additional section was added, which describes the steps needed to be
performed in order to get the appropriate kernel sources when Git is
used.

The subversion section wasn't removed, but it was marked as "legacy", as
some of the olders AMI still could use the subversion for the kernel
versioning. Finding the appropriate subversion commit on the git is
harder than simply using the SVN for downloading the kernel sources.

Signed-off-by: Michal Krawczyk <mk@semihalf.com>