linux/efa: GDR, Remove static dependency of nvidia driver
Instead of having a static dependency on the nvidia module, use
symbol_get() to dynamically import the required symbols on runtime.

As a result, the GDR enabled driver can now be loaded even when there is
no nvidia driver loaded. Note that in such cases GDR will not work.

When GDR is used, the EFA driver holds a reference on the nvidia module
and only puts it when all GPU memory regions are deallocated. Meaning
that the nvidia module cannot be unloaded while there are applications
using GPU memory through EFA.

Reviewed-by: Firas JahJah <firasj@amazon.com>
Reviewed-by: Yossi Leybovich <sleybo@amazon.com>
Signed-off-by: Gal Pressman <galpress@amazon.com>