linux/ena: Change ENI stats support check to use 'capabilities' field
Previous patch added support for 'capabilities' advertisement by ENA
device. This patch makes use of the new infrastructure to check support
for the "Network Performance Metrics" of an instance.

With this change, a failure in querying the stats becomes something
un-expected and should be treated as an error.
This patch changes the log message level for the stats query to error.

Change-Id: I521c9c9be1913721197262fcd1a8062371a950cf
Signed-off-by: Shay Agroskin <shayagr@amazon.com>