linux/ena: Move reset completion print to the reset function
The ena_restore_device() is called by the reset
function (ena_fw_reset_device()) but also from other locations in the
code. For this reason, the print which indicates that the reset
procedure has finished, should be moved from ena_restore_device() to the
reset function where it is more accurately described what happened.

Change-Id: Ic48ad4d57c74851cdc9c4ba9bb86239d13eff377
Signed-off-by: Shay Agroskin <shayagr@amazon.com>