linux/ena: Prevent LLQ initialization when membar isn't exposed
The ena_com_config_dev_mode() function contains many LLQ related
calculations and an admin command to configure the LLQ configurations in
the device.

All these operations are redundant if the driver fails to recognize the
memory bar through which LLQ is accessed (be that a HW/Linux or FW fault).

This patch first tries to map the memory bar and configures LLQ only if
it succeeds.

Another motivation for this patch is to support reloading the driver
without freeing (and reallocating) *all* its resources (what
modprobe/rmmod do). Such reload operation may allow the driver to
change LLQ entry size each time and knowing if LLQ membar is exposed in
advance would make such scenario easier to implement.

Change-Id: Ib1a3900cb0c733729165e609c3c7f6a30dd2fa9b
Signed-off-by: Shay Agroskin <shayagr@amazon.com>