linux/ena: Calculate queue parameters every device initialization
The allowed queue sizes depend on the negotiation between the driver and
the device. Since we perform this negotiation every device reset (e.g.
reset flow), it makes sense to recalculate these values every time as
well.

This would also allow to change values which affect queue parameters
between device resets. For example, it would allow the
force_large_llq_header module parameter to be taken into account every
reset.

The ena_calc_io_queue_size() function was also added with handling for
the case in which the IO queue were already configured before
ena_device_init() was called (e.g. during reset). This is needed to
avoid overriding queue sizes that were configured by the user.

As a small last change, the return value of ena_calc_io_queue_size is
changed to void since it is always 0 and is never used.

Change-Id: Ic3bff1a7fe7dc51d7c2b21a6fe8485d0b2a5e932
Signed-off-by: Shay Agroskin <shayagr@amazon.com>