linux/ena: Move large llq header parameter into ena_adapter struct
This patch moves the large_llq_header flag, which indicates if the user
requested large LLQ header (via module param) into ena_adapter struct.

This would allow this parameter to be dynamic and be changed in other
files as well.
Also disable large LLQ header configuration if the device reports that
it doesn't support it. This way the driver won't try to reconfigure it
each driver reset.

Change-Id: I7e2888666fbf10dca9f0ec9694db24ae0caac55e
Signed-off-by: Shay Agroskin <shayagr@amazon.com>