linux/ena: Postpone reset sequence if device is not responsive
Although rare, due to a possible bug, the ENA device can suffer
temporary reduction in its responsiveness. This might happen because of
unusual load on the device or bug.
This situation would usually be indicated by a MISSING_KEEP_ALIVE reset
scheduled by the driver.

If the device continues to be unresponsive throughout the reset process,
the Linux ENA driver isn't able to finish its reset routine correctly
and leaves the interface in a down state. This situation is not desired.

This patch makes the driver check whether the device is responsive
before starting the reset routine. The driver would continue trying to
reset the device until it succeeds since the reset flag is cleared only
after a reset routine is completed.

Change-Id: Ie8015ca99c885796cd514e408d5b4c0b54de14b3
Signed-off-by: Shay Agroskin <shayagr@amazon.com>