linux/efa: GDR, Fix potential lock inversion deadlock
The cleanup flow makes a couple of calls to nvidia driver with the nvmem
mutex already acquired, which can lead to a potential deadlock in case a
different flow takes the nvidia mutex first and only then the nvmem one.

In order to solve that, the nvmem_list_lock is now acquired for a
smaller section of the nvmem_put() function and the nvmem_release() call
is moved out of the critical section.

The nvidia_p2p_free_page_table() and nvidia_p2p_free_dma_mapping()
functions are no-ops and their calls have been removed.

Reviewed-by: Firas JahJah <firasj@amazon.com>
Reviewed-by: Yossi Leybovich <sleybo@amazon.com>
Signed-off-by: Gal Pressman <galpress@amazon.com>