linux/ena: Add toeplitz_calc.py description to ENA_Linux_Best_Practices.rst
The toeplitz_calc.py script, which can be found in [1], calculates the
Toeplitz hash value that the ENA HW would calculate for a given flow
(identified by a 4-tuple).

This commit adds a description of it in ENA_Linux_Best_Practices.rst
document to help users which want to fine-tune their network.

[1]: https://github.com/amzn/amzn-ec2-ena-utilities/tree/main/ena-toeplitz

Signed-off-by: Zorik Machulsky <zorik@amazon.com>
Signed-off-by: Shay Agroskin <shayagr@amazon.com>