linux/ena: Rename xdp_ring to tx_ring to describe XDP channel TX
When an XDP program is loaded the existing channels in the driver split
into two halves:
- The first half of the channels contain RX and TX rings, these queues
  are used for receiving traffic and sending packets originating from
  kernel.
- The second half of the channels contain only a TX ring. These queues
  are used for sending packets that were redirected using XDP_TX
  or XDP_REDIRECT

Referring the queues in the second half of the channels as "xdp_ring" is
confusing to the reader and might create the impression that ENA is
capable of creating an additional special-queue.

Moreover, the driver had an additional xdp_ring in ena_napi struct to
refer to the same queue tx_ring can refer to (though this problem can be
solved by putting both pointers in a union).

This patch makes the field xdp_ring be used only to describe the XDP TX
queue that a specific RX queue needs to use when forwarding packet with
XDP TX and XDP REDIRECT (i.e. leaving the xdp_ring field in ena_ring
intact).

Signed-off-by: Shay Agroskin <shayagr@amazon.com>