linux/ena: Allow setting XSK pools
The XSK pool is a contiguous memory region which a process allocates and
passes to the driver. When working with AF XDP zero-copy both TX and RX
buffers are allocated from this memory region.
Although the HW can probably deal with smaller buffers than a whole
page, for now only allow these settings.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>