linux/ena: Support polling for XSK zero-copy xmit completions
To avoid code duplication, AF-XDP zero-copy TX completions polling
function is shared with the XDP Redirect/TX one.
Most of the xmit completion handling is the same except the following
things:
- For XDP Redirect/TX, the packet is associated with an xdp_frame which
  needs to be unmapped, and freed. Since AF XDP zc receives a
  buffer from a DMA mapped user memory buffer, it isn't unmapped/freed
  and is simply acked.

- For AF-XDP zc, make the application know that it can now trigger
  another TX transmission (which would be implemented in a future
  patch) using xsk_set_tx_need_wakeup();

This patch also changes XDP napi's return code. This value is mostly
relevant for RX, unless the 'budget' value is returned in which case the
napi is called again. Given this, the completion polling function return
value is changed to boolean to indicate whether another napi cycle is
needed or not.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>