linux/ena: Add AF XDP zero-copy xmit function
The AF XDP ZC packets are transmitted as part of the napi routine. Each
packet is pointed by an xdp_desc from which the virtual and DMA address
can be extracted.

The transmission function is pretty much the same as for XDP TX/Redirect
(ena_xdp_xmit_frame()) and non-XDP case (ena_start_xmit()) but since it
uses its own data structures, much of the code has been re-written.

Also, since the napi routine for AF XDP zc will be shared with the napi
routine for XDP TX/Redirect queues in future patch, add a check for
which one is used for this specific queue (only AF XDP zc queues hold an
xsk_pool pointer).

Signed-off-by: Shay Agroskin <shayagr@amazon.com>