linux/ena: Support allocating XDP buffer from UMEM
In AF XDP zero-copy RX buffers are allocated from the UMEM (a contiguous
DMA mapped memory region allocated by and application).

It is up to the AF XDP user space application to make sure buffers are
available, and so failure to allocate buffers isn't considered an error
(a user might decide that he/she doesn't desire to receive packets and
not fill the FILL ring through which the buffers get to the driver).

The allocated xdp buffer DMA address is already added the buffer's
headroom which was configured by the application, so values like
rx_ring->headroom are completely ignored.
The buffer pointer is stored in a union with the DMA address and page
pointer in ena_rx_buffer struct since it cannot be used with them, and the
ena_rx_buffer struct should be as cache friendly as possible.

Also change the RX ring memory model to MEM_TYPE_XSK_BUFF_POOL. This is
requirement for RX queues that use buffers from UMEM.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>