linux/ena: Add AF XDP zero-copy RX polling function
The function processes descriptors which were written to the UMEM by the
device.

While the function could be used without an XDP program being loaded, it
makes little sense to do it and so the function drops any packets which
consist of several descriptors (which cannot be processed by an XDP
program).

If a packet is forwarded (either with XDP TX or Redirect) the kernel
takes care to free its descriptors. If the packet is passed (XDP_PASS)
or dropped (XDP_DROP) its memory is reused in the next descriptors passed
to the device (same as copybreak packets are "recycled").

Since the kernel doesn't allow to pass UMEM memory to the network stack,
if a packet is passed, its memory is copied to a newly allocated SKB. All
further processing of the packet is identical to the non XDP RX flow.

Similar for the TX side, the driver tells the application that it
doesn't need it to trigger another napi cycle if it's polled 'budget'
packets, or requests such a trigger otherwise. This is done using
xsk_set_rx_need_wakeup() and xsk_clear_rx_need_wakeup() functions.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>