linux/ena: Add .ndo_xsk_wakeup handler
This function can be called by an application, which has an AF XDP
zero-copy queue, to tell the driver that it requires it to poll for RX
packets, transmit enqueued packets and poll for TX completions.
In our driver it is done by scheduling the napi handler for this queue.

To avoid a case where the application busy-polls our napi handler and
causes unnecessary interrupt unmask, add a check of whether we got to
the handler by an interrupt (similar to non-XDP napi routine).

Signed-off-by: Shay Agroskin <shayagr@amazon.com>