linux/ena: Always register RX queue info
The RX queue info contains information about the RX queue which might
be relevant to the kernel.

Currently this information is only used when redirecting packets with
XDP (either to a socket or to another interface) but with the
introduction of AF XDP ZC support, the RX queue info should be
configured even if an XDP program isn't yet loaded.

This patch moves the RX queue configuration to ena_up()/ena_down()
function to always have it configured.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>