linux/ena: XDP ZC: Don't print warnings for empty RX queues
When using zero-copy AF XDP RX queues, it is up to the application to
provide the driver with RX buffers. In some cases an application might
not provide any RX buffers (for example if it only wants to handle TX
traffic) and so there's no point in triggering the driver to refill the
its RX queues automatically.

In case the application does decide to provide the driver with new RX
buffers, it has the wakeup mechanism to schedule napi and refill the
descriptors.

This patch removes the warning prints and error handling in case RX
queue buffer allocation fails for a ZC RX queue.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>