dpdk: Add kernel 5.15 support to vfio setup script
vfio-pci structure changes in the kernel 5.15 and now the vfio-pci-core
is performing memory mapping. Starting from Linux 5.15 the new patch
must be used.

Also the module installation script had to be updated in order to
install vfio-pci-core.ko module as well. As previous kernels are lacking
this sub-module, the script was updated to do not fail if there is no
such module.

Signed-off-by: Michal Krawczyk <mk@semihalf.com>