fbsd/ena: Avoid reset of the unresponsive device
If the device becomes unresponsive, the driver will not be able to
finish the reset process correctly. Timeout during version validation
indicates that the device is currently not responding. In that case
do not perform the reset and instead reschedule timer service. Because
of that the driver will continue trying to reset the device until it
succeeds or is detached.

Signed-off-by: Dawid Gorecki <dgr@semihalf.com>
Change-Id: I7f15bf42cef959e45c939cb84ff51f0785bc826c