dpdk/ena: add workaround for vfio update on Ubuntu
The Ubuntu kernels newer than 5.4.0 no longer builds the vfio kernel
driver as a module. Instead, it is being integrated into the kernel.

This changes breaks the 'get-vfio-with-wc.sh' script, which allows to
patch the vfio-pci in order to support the Write Combining. However it
is only compatible with the driver distributed as a kernel module.

The two workarounds were added in order to make the above script work
again and give the users possibility to use the vfio-pci module with
the Write Combining.

One workaround is to simply revert the kernel to the last version which
distributed the vfio-pci as a module.

The second workaround is to rebuild the whole kernel with the modified
configuration. The helper script was provided in order to automate those
steps on the Ubuntu.

Signed-off-by: Michal Krawczyk <mk@semihalf.com>
Signed-off-by: Dawid Gorecki <dgr@semihalf.com>