linux/ena: Backport bpf_warn_invalid_xdp_action() change
The bpf_warn_invalid_xdp_action() function is used by our driver after
executing an eBPF (XDP) program to signal the kernel that we received an
unsupported XDP verdict. This is either an invalid verdict or a verdict
that the current driver doesn't yet support.

Commit [1] modified the function to provide more information to the user
to help debugging such issues. This commit backports the change to
end-drivers.

[1]: https://www.spinics.net/lists/netdev/msg782898.html

Signed-off-by: Shay Agroskin <shayagr@amazon.com>