linux/ena: Replace ENA local ENA_NAPI_BUDGET to global NAPI_POLL_WEIGHT
This patch was submitted upstream for ENA linux driver in [1]

NAPI_POLL_WEIGHT is a parameter passed to the napi handling routine.
This parameter is intended to define the maximal number of packets that
a napi routine is able to process in each polling phase.

The mentioned patch is a part of a larger patchset which intends to remove
all copies of the locally defined NAPI_POLL_WEIGHT per driver.
This change aims to make refactoring easier since it removes the
locally defined field and defines a global one instead

Older kernels may not have NAPI_POLL_WEIGHT defined (for example Oracle
 Linux uek3 kernel 3.8.13). Fixed this case in kcompat.h.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git/commit/?id=36ffca1afea9

Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Osama Abboud <osamaabb@amazon.com>