linux/ena: Add a counter for driver's reset failures
Starting v2.6.0 the driver should avoid invoking VF reset procedure
if it believes that the reset wouldn't result in a functional
VF.
Nevertheless, from time to time the driver still fails to complete
the reset routine which leads into a bad state or to continuous
calls to reset.
In this patch we've added a counter to the ena_adapter struct in
order to properly keep track of reset failures.
The counter is incremented every time either ena_restore_device() or
ena_destroy_device() fail.
Also, this patch adds a print of the new stats upon running
ethtool -S.

Signed-off-by: Osama Abboud <osamaabb@amazon.com>