linux/ena: Make PHC an opt-in feature
Currently the PHC (PTP HW clock) code is included in the driver's code
if the ptp library is either compiled into the running kernel's code or
provided as a module. This can be checked using the macro
    IS_ENABLED(CONFIG_PTP_1588_CLOCK)

Otherwise, the PHC implementation is removed by the preprocessor.

In case PTP is compiled as module, a user might not have the ptp module
loaded when trying to load ENA.
If the user uses `modprobe` for that, ptp module would be loaded
automatically with ENA.
If, however, the user uses `insmod` command for that (which is more
frequently used since modprobe requires installing the driver in a defined
location) then the driver load would fail with
the error message (dmesg):

    Unknown symbol ptp_clock_index (err -2)
    Unknown symbol ptp_clock_register (err -2)
    Unknown symbol ptp_clock_unregister (err -2)

To avoid confusing users, it was decided to make PTP a compile-time
opt-in feature in ena-drivers and Github version (i.e. it is disabled
unless explicitly requested). The feature can be enabled at compile-time
by setting the envar ENA_PHC_INCLUDE. E.g.
    ENA_PHC_INCLUDE=1 make
or
    export ENA_PHC_INCLUDE=1
    make

Also update the README to help users avoid situation in which they load
the ENA driver without the ptp module.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>