fbsd/ena: Remove timer service re-arm on ena_restore_device failure
In case the reset sequence fails (ena_destroy_device() followed by
ena_restore_device() calls) during ena_restore_device(), the driver
resources are being freed. After the clean-up, the timer service is
re-armed in order to try and re-initialize the driver state.
But, such an attempt would fail given that the resources are freed.
Moreover, this would actually cause either the system to be stuck
(FreeBSD 13.1) or fail on a panic (FreeBSD 14.0).
When the driver fails in ena_restore_device() procedure, the only
recovery is either unloading and loading the driver or instance
reboot.

This change removes the timer service re-arm in case of failure
in ena_restore_device().

Fixes: 00b774b5b58a8 ("fbsd/ena: Start timer service on attach")
Signed-off-by: David Arinzon <darinzon@amazon.com>
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>