linux/ena: Fix devlink large LLQ config not fully applied
When configuring enable/disable of large LLQ via devlink
and reloading the driver, the update of tx_max_header_size
(which controls the max size to be copied from the
packet headers to the LLQ entry), is not being reflected
in the netdev TX rings. This causes a misalignment between
the netdev TX rings and the ena_com_dev.

The impacts of this misalignment are:
1. Though large LLQ is enabled, the logic will still copy the
   LLQ inline header up to the regular size of 96 bytes
   (and not 224 bytes, as expected).
   This will cause packets with longer network headers
   (such as IPv6 + TCP SACK) to be dropped on device side.
2. If large LLQ is requested via module param, and then
   disabled by devlink (set to false), traffic would
   fail, as ena_com_dev would operate with standard LLQ
   inline header max size of 96 but the netdev will
   copy up 224 bytes and the LLQ inline header will exceed
   the max size in ena_com_dev.
   The impact of this is a loop of resets, which can be
   remediated only by driver reload.

This patch sets the tx_max_header_size for the TX rings
upon running ena_restore_device.

Fixes: 09da13f46192 ("linux/ena: Add devlink support to ENA")
Signed-off-by: David Arinzon <darinzon@amazon.com>