linux/ena: Remove ifdef in ena_unmask_interrupt
The current code doesn't comply with reverse
Christmas tree format in case ENA_XDP_SUPPORT
is enabled.
Instead of having ifdefs in the logic,
initialize the rx_interval value with a
default value, and override it in case rx_ring
is not NULL (non-XDP case).

In addition, adding a space between variable
declaration and the function body.

Signed-off-by: David Arinzon <darinzon@amazon.com>