linux/ena: Fix lack of DMA sync for cpu before accessing a buffer
When swiotlb is enabled, there is no network connectivity to the instance.

It was found that when DRB is used, we do not sync the data to the CPU
side before accessing it. When DRB is not used this sync is done during
the unmapping of the dma buffer.

Even though the lack of DMA sync has yet to cause issues, syncing a
memory buffer to the CPU side after the HW owned it is necessary.
Such a sync makes sure that all the writes from the HW were flushed
into the instance memory.

With swiotlb on, the lack of DMA sync breaks the RX data path, because
swiotlb uses intermediate bounce buffers to store the RX packets that
arrive from the device. The data is copied to the ena RX queue buffers
only during sync (or unmap).

This commit adds an explicit dma sync for each RX buffer that is
recycled by DRB.

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>