linux/ena: Add a field for no interrupt moderation update action
Changes in the interrupt moderation values are infrequent from
the guest side. From gen 7 and on, the device will take the hint
set by the guest driver into account when handling interrupt
moderation, if none of the values (tx and rx) changed since
the last assignment.
This change is intended to save work for the device.

In order to keep backward compatibility, using the value '0'
to indicate that either of the interrupt moderation fields
has been updated.

Signed-off-by: David Arinzon <darinzon@amazon.com>