linux/ena: Add a hint on interrupt moderation update to the device
Changes in the interrupt moderation values for TX/RX are infrequent
from guest side. When asking to unmask interrupts, the driver always
provides the requested interrupt moderation values, regardless of
whether they've changed or not.
A new mechanism has been developed in new devices (gen 7 and on),
which allows to selectively act based on whether interrupt moderation
values have changed from the previous request or not
(and by that, saving the processing on the device side).

When a change in the interrupt moderation value is made (either by
DIM or through an ethtool command), a field is updated to reflect
this change. When asking to unmask interrupts, the driver checks
the above mentioned field to see if any of the interrupt moderation
values (TX or RX) for the particular queue whose interrupts
are now being unmasked has changed and updates the device accordingly.
The mentioned field is being reset (set to false) during the
procedure which triggers the interrupt unmask.

Signed-off-by: David Arinzon <darinzon@amazon.com>